add api and torrent tests
